requirements.txt
flask
requests
python-dotenv
spacy
huggingface_hub
PySide6
# After installing requirements, run:
# python -m spacy download en_core_web_sm